Delay configuring the node task dependencies until after project eval.
The filtering for addTasks does not work because Gradle has not
evaluated the entire script when the plugin is being applied. This moves
the tasking filtering part from the node.addTasks property into an
project.afterEvaluate block.

All the base tasks are added to the project with the same naming
convention (i.e test -> testNode if test already exists), but testNode
is only added as dependency to test if it is not filtered out by the
node.addTasks property